Create Arduino IDE code WQ
This code represents a real-time project on Water Quality Monitoring System using IoT and ML.
A research paper on this project has been published on IEEE Xplore. 
The code specifically focuses on the IoT part and is written in Arduino IDE. To execute the code, you will need to install Arduino IDE on your system and have sensors such as pH, turbidity, and TDS connected to an ESP32 module using a breadboard. Once the code is uploaded and running, it fetches data from the sensors and sends it to the cloud platform called ThingSpeak. Subsequently, the data is transmitted to a Google Colab workspace where data pre-processing and machine learning techniques are implemented. This integrated approach enables continuous monitoring of water quality parameters and facilitates the application of ML algorithms for analysis and prediction.